Added a draft version with wavelet compression of sensitivity kernel.
Notes:
(1) This version will not work as the sensitivity kernel is compressed
and no original kernel is stored to calculate the data.

(2) For parallel version, we need to modify the Haar3D(line, nx, ny, nz)
call, as we do not have nx*ny*nz elements on each CPU.

(3) The wavelet compression function needs to specify nx, ny, nz, which
we do not have for lego-grid. We need to see if it can work well with
the 1 1 N format.

(4) Currently we apply compression before the depth-weighting. While in
the Ghana paper it was done the opposite. Does it matter? If yes, we
need to pass the column-weight to sensitiviy compression, and compress
the scaled sensitivity.